refactor(client): create a base client for Lybic API
- Introduce _LybicBaseClient as the common base for Lybic API clients
- Move initialization logic to the base client- Update LybicAsyncClient and LybicClient to inherit from _LybicBaseClient
- Remove duplicate code and improve maintainability
